requirements.txt
fastapi>=0.115,<1.0
uvicorn[standard]>=0.30,<1.0
httpx>=0.27,<1.0
python-dotenv>=1.0,<2.0
rapidfuzz>=3.9,<4.0
pandas>=2.2,<3.0
Pillow>=10.0,<11.0
